Refactor SearchPhotos Component to Use Reusable PhotoGallery Component
- Updated SearchPhotos component to utilize the reusable PhotoGallery component.
- Created custom hook (useSearchPhotos) to manage fetching photos, modal state, and loading more photos with throttling.
- Enhanced code modularity and readability.
- Ensured that the photo gallery is rendered efficiently with the necessary props passed to the PhotoGallery component.
